Add ProfileFactory, change migrations

+++ database/migrations/2022_10_10_140304_create_profiles_table.php
@@ -6,49 +6,41 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
     public function up()
     {
         Schema::create('profiles', function (Blueprint $table) {
             $table->id();
-            $table->string('name');
-            $table->string('patronymic');
-            $table->string('surname');
-            $table->string('avatar');
+            $table->string('first_name');
+            $table->string('last_name')->nullable();
+            $table->string('patronymic')->nullable();
+            $table->string('slug');
+            $table->string('gender', 10)->nullable();
+            $table->string('avatar')->nullable();
             $table->date('date_birth');
-            $table->string('place_birth');
             $table->date('date_death');
+            $table->string('birth_place')->nullable();
             $table->string('burial_place')->nullable();
             $table->string('reason_death')->nullable();
             $table->string('death_certificate')->nullable();
             $table->string('religious_views')->nullable();
             $table->string('hobby')->nullable();
-            $table->integer('id_father')->nullable();
-            $table->integer('id_mother')->nullable();
-            $table->integer('id_spouse')->nullable();
-            $table->boolean('moderation_status')->nullable();
+
+            $table->string('status',16);
             $table->string('moderators_comment')->nullable();
-            $table->string('setting_access')->nullable();
-            $table->string('gender')->nullable();
+            $table->string('access')->nullable();
             $table->integer('p_id')->nullable();
             $table->integer('m_id')->nullable();
             $table->integer('f_id')->nullable();
 
+            $table->timestamp('published_at')->nullable();
             $table->timestamps();
         });
     }
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
     public function down()
     {
-        Schema::dropIfExists('profiles');
+        if (app()->isLocal()) {
+            Schema::dropIfExists('profiles');
+        }
     }
 };
